# fitness-cms/web-cms/public/swagger.yaml
openapi: 3.0.3
info:
  title: FitToday CMS API
  description: |
    API for the FitToday fitness CMS platform. Connects personal trainers with students
    through programs, workouts, and real-time communication.

    ## Authentication
    All protected endpoints require a Firebase Auth token in the Authorization header:
    ```
    Authorization: Bearer <firebase-id-token>
    ```

    ## Roles
    - **trainer** - Personal trainers who create programs and manage students
    - **student** - Students who follow programs and track progress
    - **admin** - Platform administrators

    ## Plans
    | Plan    | Price (BRL) | Commission | Max Programs | Max Students |
    |---------|-------------|------------|--------------|--------------|
    | Starter | Free        | 10%        | 3            | 5            |
    | Pro     | R$97/mo     | 5%         | Unlimited    | Unlimited    |
    | Elite   | R$197/mo    | 0%         | Unlimited    | Unlimited    |
  version: 1.0.0
  contact:
    name: FitToday Engineering
servers:
  - url: https://fittoday.vercel.app
    description: Production
  - url: http://localhost:3001
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID Token
      description: Firebase Authentication ID token
  schemas:
...
    ConnectionRequest:
      type: object
      properties:
        id:
          type: string
        trainerId:
          type: string
        studentId:
          type: string
        status:
          type: string
          enum: [pending, active, rejected, cancelled]
        source:
          type: string
        message:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        respondedAt:
          type: string
          format: date-time
          nullable: true
        cancelledAt:
          type: string
          format: date-time
          nullable: true
        cancelledBy:
          type: string
          nullable: true
          enum: [trainer, student]
        cancellationReason:
          type: string
          nullable: true
        student:
          type: object
          nullable: true
          properties:
            uid:
              type: string
            displayName:
              type: string
            email:
              type: string
            photoURL:
              type: string
              nullable: true
...
    Exercise:
      type: object
      properties:
        id:
          type: string
        name:
          type: object
          properties:
            en:
              type: string
            pt:
              type: string
        aliases:
          type: array
          items:
            type: string
        description:
          type: string
        category:
          type: string
        muscleGroups:
          type: object
          properties:
            primary:
              type: array
              items:
                type: string
            secondary:
              type: array
              items:
                type: string
        equipment:
          type: string
        force:
          type: string
          nullable: true
        level:
          type: string
          enum: [beginner, intermediate, advanced]
        mechanic:
          type: string
          nullable: true
        instructions:
          type: object
          properties:
            en:
              type: array
              items:
                type: string
            pt:
              type: array
              items:
                type: string
        media:
          type: object
          properties:
            thumbnailURL:
              type: string
            images:
              type: array
              items:
                type: string
            gifURL:
              type: string
              nullable: true
            videoURL:
              type: string
              nullable: true
        source:
          type: string
          enum: [system, trainer]
        trainerId:
          type: string
          nullable: true
        isActive:
          type: boolean
        isApproved:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateExerciseRequest:
      type: object
      required:
        - name
        - category
        - equipment
      properties:
        name:
          type: object
          required:
            - en
            - pt
          properties:
            en:
              type: string
            pt:
              type: string
        aliases:
          type: array
          items:
            type: string
        description:
          type: string
        category:
          type: string
        muscleGroups:
          type: object
          properties:
            primary:
              type: array
              items:
                type: string
            secondary:
              type: array
              items:
                type: string
        equipment:
          type: string
        force:
          type: string
        level:
          type: string
          enum: [beginner, intermediate, advanced]
          default: intermediate
        mechanic:
          type: string
        instructions:
          type: object
          properties:
            en:
              type: array
              items:
                type: string
            pt:
              type: array
              items:
                type: string
        media:
          type: object
          properties:
            thumbnailURL:
              type: string
            images:
              type: array
              items:
                type: string
            gifURL:
              type: string
            videoURL:
              type: string
        wgerBaseId:
          type: integer

    # ---- Workouts ----
...
paths:
...
  /api/connections:
    get:
      tags: [Connections]
      summary: List connection requests for trainer
      description: Trainer-only. Returns connection requests enriched with student info.
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, active, rejected, cancelled]
            default: pending
      responses:
        "200":
          description: Connection list
          content:
            application/json:
              schema:
                type: object
                properties:
                  connections:
                    type: array
                    items:
                      $ref: "#/components/schemas/ConnectionRequest"
                  total:
                    type: integer
        "401":
          description: Unauthorized
        "403":
          description: Trainer only
...
  /api/exercises/{id}:
    get:
      tags: [Exercises]
      summary: Get a single exercise
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Exercise details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"
        "404":
          description: Exercise not found

    patch:
      tags: [Exercises]
      summary: Update an exercise
      description: Trainer-only. Only the exercise owner can update trainer exercises.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExerciseRequest"
      responses:
        "200":
          description: Updated exercise
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"
        "401":
          description: Unauthorized
        "403":
          description: Not the exercise owner
        "404":
          description: Exercise not found

    delete:
      tags: [Exercises]
      summary: Soft delete exercise (set isActive=false)
      description: Trainer-only. Only the exercise owner can delete.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Exercise deactivated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        "401":
          description: Unauthorized
        "403":
          description: Not the exercise owner
        "404":
          description: Exercise not found
...